backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pymongo==4.6.0
pandas==2.1.3
numpy==1.26.2
plotly==5.18.0
python-multipart==0.0.6
cohere==4.32
pydantic==2.5.1
pydantic-settings==2.1.0
openpyxl==3.1.2
chardet==5.2.0
aiofiles==23.2.1
pdfkit==1.0.0
Jinja2==3.1.2
